requirements.txt
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-allauth==0.51.0
django-cors-headers==3.13.0
dj-rest-auth==2.2.5
drf-yasg==1.21.4
mysqlclient==2.0.3
numpy==1.23.4
pandas==1.5.1
urllib3==1.26.12
beautifulsoup4==4.11.1
django-environ==0.9.0